style: Move format_cell_t_to_string from common to format
