+++ AMBuilder
@@ -7,6 +7,7 @@ projectName = 'log4sp'
 sourceFiles = [
   'src/extension.cpp',
   'src/log4sp/common.cpp',
+  'src/log4sp/format.cpp',
   'src/log4sp/logger.cpp',
   'src/log4sp/adapter/logger_handler.cpp',
   'src/log4sp/adapter/sink_handler.cpp',
