+++ src/log4sp/common.h
@@ -133,8 +133,5 @@ private:
 
 [[nodiscard]] spdlog::filename_t unbuild_path(SourceMod::PathType type, const spdlog::filename_t &filename) noexcept;
 
-[[nodiscard]] std::string format_cell_to_string(SourcePawn::IPluginContext *ctx, const cell_t *params, const unsigned int param);
-[[nodiscard]] spdlog::memory_buf_t format_cell_to_mem_buf(SourcePawn::IPluginContext *ctx, const char *format, const cell_t *params, unsigned int *param);
-
 
 }   // namespace log4sp
